% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql_lines.R
\name{collapse_sql_lines}
\alias{collapse_sql_lines}
\title{Collapse SQL lines into single character}
\usage{
collapse_sql_lines(x)
}
\arguments{
\item{x}{character vector of SQL lines}
}
\value{
a character vector of length 1
}
\description{
A thin wrapper around \code{paste0(x, collapse = '\\\\n')} to standardize the way
qryflow collapses SQL lines.
}
\examples{
path <- example_sql_path()

lines <- read_sql_lines(path)

sql <- collapse_sql_lines(lines)
}
