% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_sql_path}
\alias{example_sql_path}
\title{Get path to qryflow example SQL scripts}
\usage{
example_sql_path(path = "mtcars.sql")
}
\arguments{
\item{path}{filename of the example script.}
}
\value{
path to example SQL script
}
\description{
qryflow provides example SQL scripts in its \code{inst/sql} directory. Use this
function to retrieve the path to an example script. This function is intended
to facilitate examples, vignettes, and package tests.
}
\examples{
path <- example_sql_path("mtcars.sql")

file.exists(path)
}
