% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R, R/parsers.R, R/type-registration.R
\name{ls_qryflow_handlers}
\alias{ls_qryflow_handlers}
\alias{ls_qryflow_parsers}
\alias{ls_qryflow_types}
\title{List currently registered chunk types}
\usage{
ls_qryflow_handlers()

ls_qryflow_parsers()

ls_qryflow_types()
}
\value{
Character vector of registered chunk types
}
\description{
Helper function to access the names of the currently registered chunk types.
Functions available for accessing just the parsers or just the handlers.
}
\details{
\code{ls_qryflow_types} is implemented to return the union of the results of
\code{ls_qryflow_parsers} and \code{ls_qryflow_handlers}. It's expected that a both
a parser and a handler exist for each type. If this assumption is violated,
the \code{ls_qryflow_types} may suggest otherwise.
}
\examples{
ls_qryflow_types()
}
