% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qryflow_chunk.R
\name{new_qryflow_chunk}
\alias{new_qryflow_chunk}
\title{Create an instance of the \code{qryflow_chunk} class}
\usage{
new_qryflow_chunk(
  type = character(),
  name = character(),
  sql = character(),
  tags = NULL,
  results = NULL
)
}
\arguments{
\item{type}{Character indicating the type of chunk (e.g., "query", "exec")}

\item{name}{Name of the chunk}

\item{sql}{SQL statement associated with chunk}

\item{tags}{Optional, additional tags included in chunk}

\item{results}{Optional, filled in after chunk execution}
}
\value{
An list-like object of class \code{qryflow_chunk}
}
\description{
Create an instance of the \code{qryflow_chunk} class
}
\details{
Exported for users intending to extend qryflow. Subsequent processes rely on
the structure of a qryflow_chunk.
}
\examples{
chunk <- new_qryflow_chunk("query", "df_name", "SELECT * FROM mtcars;")
}
