% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qryflow.R
\name{qryflow_default_type}
\alias{qryflow_default_type}
\title{Access the default qryflow chunk type}
\usage{
qryflow_default_type(type = getOption("qryflow.default.type", "query"))
}
\arguments{
\item{type}{Optional. The type you want to return.}
}
\value{
Character. If set, result from \code{qryflow.default.type} option, otherwise "query" or value passed to \code{type}
}
\description{
Retrieves the value from the option \code{qryflow.default.type}, if set. Otherwise returns
"query", which is the officially supported default type. If any value is supplied
to the function, it returns that value.
}
\examples{
x <- getOption("qryflow.default.type", "query")

y <- qryflow_default_type()

identical(x, y)
}
