% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{qryflow_handler_exists}
\alias{qryflow_handler_exists}
\title{Check existence of a given handler in the registry}
\usage{
qryflow_handler_exists(type)
}
\arguments{
\item{type}{chunk type to check (e.g., "query", "exec")}
}
\value{
Logical. Does \code{type} exist in the handler registry?
}
\description{
Checks whether the specified handler exists in the handler registry environment.
}
\examples{
qryflow_handler_exists("query")
}
\seealso{
\code{\link[=qryflow_parser_exists]{qryflow_parser_exists()}} for the parser equivalent.
}
