% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{qryflow_meta}
\alias{qryflow_meta}
\title{Extract metadata from qryflow objects}
\usage{
qryflow_meta(x)
}
\arguments{
\item{x}{\code{qryflow} or \code{qryflow_chunk} object}
}
\description{
Extract metadata from qryflow objects
}
\examples{
con <- example_db_connect(mtcars)

filepath <- example_sql_path("mtcars.sql")

parsed <- qryflow_parse(filepath)
qryflow_meta(parsed)
qryflow_meta(parsed[[1]])

results <- qryflow_execute(con, parsed)
qryflow_meta(results)
qryflow_meta(results[[1]])

DBI::dbDisconnect(con)
}
