% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{qryflow_parser_exists}
\alias{qryflow_parser_exists}
\title{Check existence of a given parser in the registry}
\usage{
qryflow_parser_exists(type)
}
\arguments{
\item{type}{chunk type to check (e.g., "query", "exec")}
}
\value{
Logical. Does \code{type} exist in the parser registry?
}
\description{
Checks whether the specified parser exists in the parser registry environment.
}
\examples{
qryflow_parser_exists("query")
}
\seealso{
\code{\link[=qryflow_handler_exists]{qryflow_handler_exists()}} for the handler equivalent.
}
