% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handlers.R
\name{validate_qryflow_handler}
\alias{validate_qryflow_handler}
\title{Ensure correct handler structure}
\usage{
validate_qryflow_handler(handler)
}
\arguments{
\item{handler}{object to check}
}
\value{
Logical. Generates an error if the object does not pass all the
criteria.
}
\description{
This function checks that the passed object is a function and contains
the arguments "con", "chunk", and "..." - in that order. This is to help ensure users
only register valid handlers.
}
\examples{
custom_func <- function(con, chunk, ...){

  # Parsing Code Goes Here

}

validate_qryflow_handler(custom_func)

}
