% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{validate_qryflow_parser}
\alias{validate_qryflow_parser}
\title{Ensure correct parser structure}
\usage{
validate_qryflow_parser(parser)
}
\arguments{
\item{parser}{object to check}
}
\value{
Logical. Generates an error if the object does not pass all the
criteria.
}
\description{
This function checks that the passed object is a function and contains
the arguments "x" and "..." - in that order. This is to help ensure users
only register valid parsers.
}
\examples{
custom_func <- function(x, ...){

  # Parsing Code Goes Here

}
validate_qryflow_parser(custom_func)
}
\seealso{
\code{\link[=validate_qryflow_handler]{validate_qryflow_handler()}} for the handler equivalent.
}
