% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-registration.R
\name{ls_qryflow_types}
\alias{ls_qryflow_types}
\title{List currently registered chunk types}
\usage{
ls_qryflow_types()
}
\value{
Character vector of registered chunk types
}
\description{
Helper function to access the names of the currently registered chunk types.
}
\examples{
ls_qryflow_types()
}
