% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-registration.R
\name{register_qryflow_type}
\alias{register_qryflow_type}
\title{Register custom chunk types}
\usage{
register_qryflow_type(type, handler, overwrite = FALSE)
}
\arguments{
\item{type}{Character indicating the chunk type (e.g., "exec", "query")}

\item{handler}{A function to execute the SQL associated with the type. Must accept arguments "chunk", "con", and "...".}

\item{overwrite}{Logical. Overwrite existing handler, if exists?}
}
\value{
Logical. Indicating whether types were successfully registered.
}
\description{
Use this function to register a custom chunk type with qryflow
}
\details{
To avoid manually registering your custom type each session, consider adding
the registration code to your \code{.Rprofile} or creating a package that leverages
\code{\link[=.onLoad]{.onLoad()}}
}
\examples{
# Create custom handler #####
custom_handler <- function(con, chunk, ...){
  # Custom execution code will go here...
  # return(result)
}

register_qryflow_type("query-send", custom_handler, overwrite = TRUE)

}
